+++ Block/Adminhtml/Entity/Grid/Renderer/Image.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+
+namespace Alekseon\AlekseonEav\Block\Adminhtml\Entity\Grid\Renderer;
+
+/**
+ * Class Image
+ * @package Alekseon\AleksoenEav\Block\Adminhtml\Entity\Grid\Renderer
+ */
+class Image extends \Magento\Backend\Block\Widget\Grid\Column\Renderer\AbstractRenderer
+{
+    /**
+     * @var \Magento\Store\Model\StoreManagerInterface
+     */
+    private $storeManager;
+
+    /**
+     * Image constructor.
+     * @param \Magento\Backend\Block\Context $context
+     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
+     * @param array $data
+     */
+    public function __construct(
+        \Magento\Backend\Block\Context $context,
+        \Magento\Store\Model\StoreManagerInterface $storeManager,
+        array $data = []
+    ) {
+        $this->storeManager = $storeManager;
+        parent::__construct($context, $data);
+    }
+
+    /**
+     * @param \Magento\Framework\DataObject $row
+     * @return string|void
+     */
+    public function render(\Magento\Framework\DataObject $row)
+    {
+        if (!$row->getData($this->getColumn()->getIndex())) {
+            return null;
+        }
+
+        $mediaUrl = $this->storeManager->getStore($row->getStoreId())
+            ->getBaseUrl(
+                \Magento\Framework\UrlInterface::URL_TYPE_MEDIA
+            );
+
+        $imagePath = $mediaUrl .  $row->getData($this->getColumn()->getIndex());
+
+        return '<img width="100" height="100" src="' . $imagePath . '"/>';
+    }
+}